requirements.txt
wheel
numpy
matplotlib
pillow
pydantic
pytest
./libs/pyutils
./libs/tt-pytorch
./libs/shnetutil
scipy
sklearn
scikit-image
joblib
numba
jupyter
requests
pycoshrem
skorch
ggplot
pydantic
torchvision
https://github.com/stefanloock/pyshearlab/archive/master.zip
pympler
cplxmodule
tensorly
pthflops